dyslexia/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.5.0
numpy>=1.19.2
pillow>=8.2.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
pandas>=1.3.0
tqdm>=4.62.0
jupyter>=1.0.0
notebook>=6.4.0 